move mapUtils to src and update paths
