+++ ROADMAP.md
@@ -114,7 +114,7 @@
 98. Страница `/tasks` позволяет скачать таблицу задач в CSV или JSON
 99. Кнопки действий задачи закреплены внизу внутри формы, цвет «Принять» и «Выполнено» зависит от статуса
 100. Команда `/task_form_app` отправляет ссылку на форму создания задачи
-101. Функции работы с Google Maps вынесены в общий модуль `bot/shared/mapUtils.js`
+101. Функции работы с Google Maps вынесены в общий модуль `bot/src/shared/mapUtils.js`
 102. Все изменения задач фиксируются в коллекции логов
 103. `authFetch` перенаправляет на `/login` при отсутствии токена или статусе 401/403
 104. Срок действия JWT увеличен до семи дней
@@ -260,7 +260,7 @@
 229. Удалён дублирующийся файл .husky/husky.sh
 230. Удалён пакет @aws-sdk/client-s3
 231. Повторяющийся код расчёта маршрута вынесен в функцию applyRouteInfo
-232. Веб‑клиент импортирует mapUtils через default объект
+232. Веб‑клиент импортирует mapUtils из `src/shared/mapUtils.js`
 233. authFetch поддерживает опцию noRedirect, предотвращая цикл входа
 234. API использует middleware compression для сжатия
 235. Фильтры логов получили атрибуты aria-label для улучшения доступности
