+++ bot/src/shared/mapUtils.js
