+++ bot/web/src/utils/createRouteLink.js
@@ -1,5 +1,5 @@
 // Формирование ссылки маршрута Google Maps из координат
-import mapUtils from '../../../shared/mapUtils.js';
+import mapUtils from '../../../src/shared/mapUtils.js';
 
 export default function createRouteLink(start, end, mode = 'driving') {
   return mapUtils.generateRouteLink(start, end, mode)
