+++ bot/web/src/utils/extractCoords.js
@@ -1,5 +1,5 @@
 // Извлекает координаты из ссылки Google Maps
-import mapUtils from '../../../shared/mapUtils.js';
+import mapUtils from '../../../src/shared/mapUtils.js';
 
 export default function extractCoords(url) {
   return mapUtils.extractCoords(url)
