+++ docs/technical_manual.md
@@ -160,7 +160,7 @@ Middleware `checkRole` и `checkTaskAccess` записывают отказ до
 4. Мини‑приложение можно открыть по команде `/task_form_app`.
 
 Бот разворачивает короткие ссылки Google Maps и сохраняет координаты задачи. После создания задачи из мини‑приложения отправляется событие `task_created` через `Telegram.WebApp.sendData`.
-Общие функции формирования ссылок и извлечения координат вынесены в `bot/shared/mapUtils.js` и используются сервером и клиентом. Веб‑клиент импортирует модуль как `import mapUtils from '../../../shared/mapUtils.js'`.
+Общие функции формирования ссылок и извлечения координат находятся в `bot/src/shared/mapUtils.js` и используются сервером и клиентом. Веб‑клиент импортирует модуль как `import mapUtils from '../../../src/shared/mapUtils.js'`.
 
 ## Проверка initData WebApp
